Export encrypt functions and fix lint errors
