+++ encrypt_file/Cargo.toml
@@ -2,6 +2,8 @@
 name = "encrypt_file"
 version = "0.1.0"
 edition = "2021"
+description = "A Rust library for encrypting files using the XChaCha20 algorithm."
+license = "MIT"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
