+++ package.json
@@ -66,11 +66,11 @@
   "homepage": "https://github.com/parajbs-dev/s5client-js",
   "dependencies": {
     "@babel/runtime": "^7.22.5",
+    "@c4312/blake3-internal": "^3.0.0",
     "@noble/hashes": "^1.3.1",
     "@types/file-saver": "^2.0.5",
     "axios": "^1.4.0",
     "base32-encode": "^2.0.0",
-    "@c4312/blake3-internal": "^3.0.0",
     "blake3-wasm": "^3.0.0",
     "buffer": "^6.0.3",
     "file-saver": "^2.0.5",
@@ -106,7 +106,7 @@
     "babel-loader": "^9.1.2",
     "copy-webpack-plugin": "^11.0.0",
     "crypto-browserify": "3.12.0",
-    "eslint": "^8.42.0",
+    "eslint": "^8.43.0",
     "eslint-plugin-jsdoc": "46.2.6",
     "lint-staged": "^13.2.2",
     "prettier": "^2.8.8",
